+++ src/components/quiz/quiz/mode/Connect.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Connect() {
+  return <div></div>;
+}
