+++ src/components/quiz/quiz/mode/Picture.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Picture() {
+  return <div></div>;
+}
